// src/app/components/assets/assetsGrid/types.ts
export interface AssetsMetadata {
    context: Context;
    taxonomy: Taxonomy;
    creators: Creators;
    provenance: Provenance;
}
export interface Context {
    formData: FormData;
    schema: Schema;
    uiSchema: UiSchema;
}
export interface FormData {
    colors?: string[] | null;
    title?: string;
}
export interface Schema {
    type: string;
    required?: string[] | null;
    properties: Properties;
}
// ...
export interface ItemsOrCultureOrOrientationOrObjectTypeOrAiGenerationOrArenabledOrNudityOrCategoryOrEthnicityOrGenderOrBlockchain {
    type: string;
    enum?: string[] | null;
}
export interface MoodOrMediumOrStyle {
    type: string;
    minItems: number;
    uniqueItems: boolean;
    items: ItemsOrCultureOrOrientationOrObjectTypeOrAiGenerationOrArenabledOrNudityOrCategoryOrEthnicityOrGenderOrBlockchain;
}
export interface ColorsOrTagsOrCollectionsOrSubjectOrRoles {
    type: string;
    item: string;
    minItems: number;
    maxItems: number;
    items: ItemsOrTitleOrCopyrightOrNameOrExhibitionNameOrAwardNameOrPlusCode;
}
export interface UiSchema {
    title: TitleOrDescriptionOrCopyrightOrNameOrBioOrProfileUrlOrNationalityOrResidenceOrEthnicityOrGenderOrCountryOrPlusCodeOrBlockchain;
    description: TitleOrDescriptionOrCopyrightOrNameOrBioOrProfileUrlOrNationalityOrResidenceOrEthnicityOrGenderOrCountryOrPlusCodeOrBlockchain;
    culture: CultureOrMoodOrOrientationOrObjectTypeOrAiGenerationOrArenabledOrNudityOrCategoryOrMediumOrStyle;
    mood: CultureOrMoodOrOrientationOrObjectTypeOrAiGenerationOrArenabledOrNudityOrCategoryOrMediumOrStyle;
    colors: Colors;
    copyright: TitleOrDescriptionOrCopyrightOrNameOrBioOrProfileUrlOrNationalityOrResidenceOrEthnicityOrGenderOrCountryOrPlusCodeOrBlockchain;
    orientation: CultureOrMoodOrOrientationOrObjectTypeOrAiGenerationOrArenabledOrNudityOrCategoryOrMediumOrStyle;
}
// ...
export interface Taxonomy {
    formData: FormData1;
    schema: Schema1;
    uiSchema: UiSchema1;
}
export interface FormData1 {
    subject?: string[] | null;
    tags?: string[] | null;
}
export interface Schema1 {
    type: string;
    required?: string[] | null;
    properties: Properties1;
}
// ...
export interface UiSchema1 {
    objectType: CultureOrMoodOrOrientationOrObjectTypeOrAiGenerationOrArenabledOrNudityOrCategoryOrMediumOrStyle;
    tags: TagsOrCollectionsOrSubject;
    collections: TagsOrCollectionsOrSubject;
    aiGeneration: CultureOrMoodOrOrientationOrObjectTypeOrAiGenerationOrArenabledOrNudityOrCategoryOrMediumOrStyle;
    arenabled: CultureOrMoodOrOrientationOrObjectTypeOrAiGenerationOrArenabledOrNudityOrCategoryOrMediumOrStyle;
    nudity: CultureOrMoodOrOrientationOrObjectTypeOrAiGenerationOrArenabledOrNudityOrCategoryOrMediumOrStyle;
    category: CultureOrMoodOrOrientationOrObjectTypeOrAiGenerationOrArenabledOrNudityOrCategoryOrMediumOrStyle;
    medium: CultureOrMoodOrOrientationOrObjectTypeOrAiGenerationOrArenabledOrNudityOrCategoryOrMediumOrStyle;
    style: CultureOrMoodOrOrientationOrObjectTypeOrAiGenerationOrArenabledOrNudityOrCategoryOrMediumOrStyle;
    subject: TagsOrCollectionsOrSubject;
}
export interface TagsOrCollectionsOrSubject {
    ['ui:widget']: string;
    items: Items1;
    ['ui:options']: Uioptions1;
}
export interface Items1 {
    ['ui:emptyValue']: string;
    ['ui:options']: Uioptions2;
}
export interface Creators {
    formData?: FormDataEntity[] | null;
    schema: Schema2;
    uiSchema: UiSchema2;
}
export interface FormDataEntity {
    name: string;
}
export interface Schema2 {
    type: string;
    items: Items2;
}
// ...
export interface UiSchema2 {
    ['ui:options']: Uioptions3;
    items: Items3;
}
// ...
export interface Provenance {
    formData: FormData2;
    schema: Schema3;
    uiSchema: UiSchema3;
}
export interface FormData2 {
    awards?: null[] | null;
    exhibitions?: null[] | null;
    blockchain: string;
}
export interface Schema3 {
    type: string;
    required?: string[] | null;
    properties: Properties3;
}
// ...
export interface UiSchema3 {
    country: TitleOrDescriptionOrCopyrightOrNameOrBioOrProfileUrlOrNationalityOrResidenceOrEthnicityOrGenderOrCountryOrPlusCodeOrBlockchain;
    plusCode: TitleOrDescriptionOrCopyrightOrNameOrBioOrProfileUrlOrNationalityOrResidenceOrEthnicityOrGenderOrCountryOrPlusCodeOrBlockchain;
    blockchain: TitleOrDescriptionOrCopyrightOrNameOrBioOrProfileUrlOrNationalityOrResidenceOrEthnicityOrGenderOrCountryOrPlusCodeOrBlockchain;
    exhibitions: Exhibitions1;
    awards: Awards1;
}